\name{BAGS-internal}
\alias{GrpMean}

\title{
Internal functions of the BAGS package.
}
\description{
These are internal functions of the package BAGS.
}
%\usage{
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%
%}
\details{
These functions perform internal tasks an work within the package BAGS.
}
%\value{
%An entry of the vector with the posterior sample for the parameter Alpha for the cross-sectional data setting.
%}
%%\references{
%% ~put references to the literature/web site here ~
%%}
\author{
A. Quiroz-Zarate.
\email{aquiroz@jimmy.harvard.edu}
}
\keyword{internal}
%%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
