\name{GeneGroupAnalysis-internal}
\alias{Alfa.Aux.cs}
\alias{Alfa0NPCFun.ts}
\alias{Alfa.Aux.cs}
\alias{Alfa0NPCFun.ts}
\alias{AlfaAllCFun.ts}
\alias{AlfaFun.cs}
\alias{AlfaNPCFun.ts}
\alias{Beta.Aux.cs}
\alias{BetaCFun.ts}
\alias{BetaFun.cs}
\alias{GrpMeanDataFun}
\alias{MakeAffyOrderGrp.Fun}
\alias{MeanMatGrps.Aux.ts}
\alias{PiFun}
\alias{RhoFun}
\alias{SgmAlfaFun.cs}
\alias{SgmAlfaFun.ts}
\alias{SgmAprGrps.Aux.ts}
\alias{SgmY.Aux.cs}
\alias{SgmYCFun.ts}
\alias{SgmYFun.cs}

\title{
Internal functions of the GeneGroupAnalysis package.
}
\description{
These are internal functions of the package GeneGroupAnalysis.
}
%\usage{
%Alfa.Aux.cs(grp.sz.i, y.mu, param, sgm.alfa.i, sgm.y.i, rho, mm)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%
%}
\details{
These functions work within the package GeneGroupAnalysis.
}
%\value{
%An entry of the vector with the posterior sample for the parameter Alpha for the cross-sectional data setting.
%}
%%\references{
%% ~put references to the literature/web site here ~
%%}
\author{
A. Quiroz-Zarate and John Quackenbush.
}

\keyword{internal}
%%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
