\name{GeneGroupAnalysis-package}
\alias{GeneGroupAnalysis-package}
\alias{GeneGroupAnalysis}
\docType{package}
\title{
The Gene Group Analysis package for Gene Functional Class Analysis.
}
\description{
Functions to perform gene-set significance analysis method for data sets with cross-sectional or time series designs.
}
\details{
\tabular{ll}{
Package: \tab GeneGroupAnalysis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-05\cr
License: \tab Artistics-2.0\cr
LazyLoad: \tab yes\cr
}
For detailed examples on the use and manipulation of the functions please see the package Vignette.
}
\author{
A. Quiroz-Zarate and John Quackenbush
\email{aquiroz@hsph.harvard.edu}
}
\references{
Quiroz-Zarate A and Quackenbush J. "Manuscript in preparation. 2012."
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{Gene Ontology}
\keyword{Gene-set significance analysis}
\keyword{Gibb's sampler}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%%\examples{
%%~~ simple examples of the most important functions ~~
%%}
