\name{BAGS-package}
\alias{BAGS-package}
\alias{BAGS}
\docType{package}
\title{
A Bayesian Approach for Geneset Selection (BAGS).
}
\description{
Functions to perform geneset significance analysis for data sets with cross-sectional or time course data design. This method is described in detail in the paper soon to be published. This work is done by Alejandro Quiroz-Zarate in collaboration with Benjamin Haibe-Kains and John Quackenbush.
}
\details{
\tabular{ll}{
Package: \tab BAGS\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2013-06-12\cr
License: \tab Artistics-2.0\cr
LazyLoad: \tab yes \cr
}
For a detailed example on the use and manipulation of the functions provided on this package please see the package Vignette.
}
\author{
A. Quiroz-Zarate
\email{aquiroz@jimmy.harvard.edu}
}
\references{
Quiroz-Zarate, A., Haibe-Kains, B. and Quackenbush, J. Manuscript in preparation (2013).
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
\keyword{Genesets}
\keyword{Enrichment}
\keyword{MSigDB}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
#vignette("BAGS")
}